Factor out test utilities in a common base class
